Create Tron_LineByTurns
This version of the Tron game takes a different approach to tracking each bike's line. This one uses a linked list of structs that keep track of every point the bike turned at. At each loop, the program uses a bunch of if statements to see if the bike's current location is in between any of those turn points.

This program is still in the works. Currently, I am trying to solve a bug where the program isn't keeping track of the turn points in order